+++ .storybook/stories/PasswordInput.stories.tsx
@@ -1,16 +1,15 @@
-import type { Meta, StoryObj } from '@storybook/react';
-import { PasswordInput } from '../../components/uikit/Input/PasswordInput';
-
+import type { Meta, StoryObj } from "@storybook/react";
+import { PasswordInput } from "../../components/uikit/Input/dumb/PasswordInput";
 
 const PasswordInputMeta: Meta<typeof PasswordInput> = {
-    title: 'Uikit/Input/PasswordInput',
-    component: PasswordInput,
-    args: {
-        isRequired: true,
-        error: "That field is required"
-    },
-  };
+  title: "Uikit/Input/PasswordInput",
+  component: PasswordInput,
+  args: {
+    isRequired: true,
+    error: "That field is required",
+  },
+};
 
-  export default PasswordInputMeta;
+export default PasswordInputMeta;
 
 export const Basic: StoryObj<typeof PasswordInput> = {};
