+++ .storybook/stories/TextInput.stories.tsx
@@ -1,17 +1,17 @@
-import type { Meta, StoryObj } from '@storybook/react';
-import { TextInput } from '../../components/uikit/Input/TextInput';
-import { User } from '../../components/icons/user';
+import type { Meta, StoryObj } from "@storybook/react";
+import { TextInput } from "../../components/uikit/Input/dumb/TextInput";
+import { User } from "../../components/icons/user";
 
 const TextInputMeta: Meta<typeof TextInput> = {
-    title: 'Uikit/Input/TextInput',
-    component: TextInput,
-    args: {
-        placeholder: 'Enter text...',
-        icon: User,
-        error: "Error here"
-    },
-  };
+  title: "Uikit/Input/TextInput",
+  component: TextInput,
+  args: {
+    placeholder: "Enter text...",
+    icon: User,
+    error: "Error here",
+  },
+};
 
-  export default TextInputMeta;
+export default TextInputMeta;
 
 export const Basic: StoryObj<typeof TextInput> = {};
