+++ components/uikit/Input/TextareaInput.tsx
@@ -1,77 +1,24 @@
+import { useField } from "formik";
 import {
-  AlertCircleIcon,
-  Box,
-  FormControl,
-  FormControlError,
-  FormControlErrorIcon,
-  FormControlErrorText,
-  FormControlHelper,
-  FormControlHelperText,
-  Textarea,
-  TextareaInput,
-} from "@gluestack-ui/themed";
+  TextAreaInput as TextAreaInputDumb,
+  TextareaInputProps as TextareaInputDumbProps,
+} from "./dumb/TextareaInput";
 
-export type TextareaInputPropos = {
-  label: string,
-  placeholder: string;
-  error?: string;
-  help?: string;
-  value?: string;
-  onChange?: (value: string) => void;
-  isDisabled: boolean;
-  isRequired: boolean;
-};
+interface TextareaInputProps
+  extends Omit<TextareaInputDumbProps, "value" | "onChange" | "error"> {
+  name: string;
+}
 
-export const TextAreaInput = ({
-  label,
-  placeholder,
-  error,
-  help,
-  value,
-  onChange,
-  isDisabled,
-  isRequired,
-}: TextareaInputPropos) => {
-  const isInvalid = error ? true : false;
+const TextareaInput = ({ name, ...props }: TextareaInputProps) => {
+  const [{ value }, { error }, { setValue }] = useField<string>(name);
   return (
-    <Box>
-      <FormControl
-        isInvalid={isInvalid}
-        isRequired={isRequired}
-        isDisabled={isDisabled}
-      >
-        <Textarea
-          p="$4"
-          h={176}
-          borderRadius="$xl"
-          borderWidth={1.5}
-          borderColor="$white"
-          $focus-borderColor="$primaryDark"
-        >
-          <TextareaInput
-            placeholder={placeholder}
-            fontSize={15}
-            p="$0"
-            placeholderTextColor="$primaryNormal"
-            color="$primaryNormal"
-            value={value}
-            onChangeText={onChange}
-          />
-        </Textarea>
-        {help ? (
-          <FormControlHelper>
-            <FormControlHelperText fontFamily="Livvic" fontSize="$sm" color="$black">
-              {help}
-            </FormControlHelperText>
-          </FormControlHelper>
-        ) : null}
-        {isInvalid ? (
-          <FormControlError>
-            <FormControlErrorIcon as={AlertCircleIcon} color="$red"/>
-            <FormControlErrorText fontSize="$sm">{error}</FormControlErrorText>
-          </FormControlError>
-        ) : null}
-      </FormControl>
-    </Box>
+    <TextAreaInputDumb
+      error={error}
+      value={value}
+      onChange={setValue}
+      {...props}
+    />
   );
 };
+
+export default TextareaInput;
