+++ package.json
@@ -39,6 +39,7 @@
     "expo-system-ui": "~3.0.4",
     "expo-web-browser": "~13.0.3",
     "firebase": "^10.12.0",
+    "formik": "^2.4.6",
     "lodash": "^4.17.21",
     "react": "18.2.0",
     "react-dom": "18.2.0",
